+++ .gitignore
@@ -151,3 +151,6 @@ dist
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# secrets files
+parksAPIKey.txt
