+++ src/App.js
@@ -1,4 +1,4 @@
-import MainNavPage from './MainNavPage';
+import MainNavPage from './components/MainNavPage';
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { fas } from '@fortawesome/free-solid-svg-icons'
 import { far } from '@fortawesome/free-regular-svg-icons'
