+++ src/components/ParksList.js
@@ -1,0 +1,1 @@
+// List of parks being displayed to user
