+++ src/components/ParksSearchSection.js
@@ -1,0 +1,4 @@
+// parent component to SearchBar and ParksSearchSection to hold state of what is being searched.
+// componentDidMount() {
+//     api call to populate this.state.parks
+// }
