+++ src/index.js
@@ -2,6 +2,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import searchParks from './api';
+
+// searchParks();
 
 // 2) Get a reference to the div with ID root
 const el = document.getElementById('root');
